BN v.0.1
- Placement des bateux (seulement Porte-Avion en test)
- Affichage de la grille

Enfin un commit dans cette branche
